moved requires to top of file. Also renamed content.js to contact.js - typo
